requirements.txt
accelerate>=1.2.0
datasets>=3.2.0
fastapi>=0.115.0
httpx>=0.28.0
peft>=0.14.0
python-dotenv>=1.0.1
python-telegram-bot>=22.1
safetensors>=0.5.2
torch>=2.5.0
transformers>=4.46.0
trl>=0.12.0
uvicorn>=0.34.0
bitsandbytes>=0.45.0

